+++ packages/layout/src/components/cards/common/views/preview/components/toolbar/CardMainHeadingTitle.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { UnderlinedButton } from '../../../../../components/button';
+
+export const CardMainHeadingTitle: React.FC<{ title: string }> = ({
+	title,
+}) => <UnderlinedButton isMainHeading={true}>{title}</UnderlinedButton>;
